+++ src/modern-async.mjs
@@ -14,8 +14,6 @@ export { default as find } from './find.mjs'
 export { default as findIndex } from './findIndex.mjs'
 export { default as findIndexLimit } from './findIndexLimit.mjs'
 export { default as findIndexSeries } from './findIndexSeries.mjs'
-export { default as findLimit } from './findLimit.mjs'
-export { default as findSeries } from './findSeries.mjs'
 export { default as forEach } from './forEach.mjs'
 export { default as forEachLimit } from './forEachLimit.mjs'
 export { default as forEachSeries } from './forEachSeries.mjs'
